LayoutTests/svg/dynamic-updates/SVGFEMorphologyElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 44: TypeError: 'undefined' is not an object (evaluating 'morphologyElement.in1.baseVal = "SourceGraphic"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Erode radius = 4
Tests dynamic updates of the 'in' property of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL morphologyElement.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'morphologyElement.in1.baseVal')
